pyproject.toml
[project]
name = "agentic-trading-system"
version = "1.0.0"
description = "Advanced agentic trading system with multi-agent architecture"
authors = [
    {name = "Trading System Team", email = "team@trading-system.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    # Core FastAPI and async
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "requests>=2.31.0",
    # Database and caching
    "redis[hiredis]>=5.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    # Data processing and analysis
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "ta-lib>=0.4.28",
    "python-dateutil>=2.8.0",
    # Blockchain and Web3
    "web3>=6.12.0",
    "eth-account>=0.9.0",
    "cryptography>=41.0.0",
    # AI and LLM
    "openai>=1.3.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    # Monitoring and observability
    "prometheus-client>=0.19.0",
    "sentry-sdk[fastapi]>=1.38.0",
    # Authentication and security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "autogen-agentchat>=0.2.0",
    "langchain-community>=0.0.10",
    # CAMEL-AI Framework
    "camel-ai[all,tools]>=0.1.0",
    # Vector Database and Embeddings
    "qdrant-client>=1.7.0",
    "llm-guard>=0.3.0",
    # "agentic-radar>=0.14.0",  # Not available on PyPI
    # "garak>=2.0.0",  # Version conflict
    # "rebuff>=0.1.0",  # Not available on PyPI
    "logfire>=0.6.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
    "fakeredis>=2.20.0",
    "pytest-xdist>=3.3.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    "requests-oauthlib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
    "fakeredis>=2.20.0",
    "pytest-xdist>=3.3.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "api*", "core*", "dex-simulator*"]
exclude = ["tests*", "data*", "logs*", "config*", "scripts*"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
    "fakeredis>=2.20.0",
    "pytest-xdist>=3.3.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ['py311']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["agents", "api", "core"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "ta.*",
    "web3.*",
    "eth_account.*",
    "langchain.*",
    "openai.*",
    "prometheus_client.*",
    "sentry_sdk.*",
    "autogen.*",
    "llm_guard.*",
    "agentic_radar.*",
    "garak.*",
    "rebuff.*",
    "camel.*",
    "qdrant_client.*",
    "sentence_transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "functional: marks tests as functional tests",
]
